requirements.txt
python-dotenv
pandas
numpy
certifi
pymongo

-e . 